split up tmy.py. split tmy_cli test into cli and regression test
