+++ nsrdb/tmy/tmy.py
@@ -1,397 +1,22 @@
-"""NSRDB Typical Meteorological Year (TMY) code.
-
-Created on Wed Oct 23 10:55:23 2019
-
-@author: gbuster
-
-TODO: Refactor using new CLI. Don't need the hpc wrapper functions in here
-"""
+"""NSRDB Typical Meteorological Year (TMY) code."""
 
 import datetime
 import logging
 import os
-import shutil
-from concurrent.futures import as_completed
 from copy import deepcopy
 from itertools import groupby
-from typing import ClassVar
 
-import h5py
 import numpy as np
 import pandas as pd
 from cloud_fs import FileSystem
-from rex import init_logger
-from rex.utilities.execution import SpawnProcessPool
 
-from nsrdb.data_model.variable_factory import VarFactory
-from nsrdb.file_handlers.outputs import Outputs
 from nsrdb.file_handlers.resource import MultiFileResource, Resource
+from nsrdb.tmy.cdf import Cdf
 from nsrdb.utilities.file_utils import pd_date_range
 
+from .utilities import drop_leap, make_time_masks
+
 logger = logging.getLogger(__name__)
-
-
-class Cdf:
-    """Class for handling cumulative distribution function methods for TMY."""
-
-    def __init__(self, my_arr, my_time_index, time_masks=None):
-        """
-        Parameters
-        ----------
-        my_arr : np.ndarray
-            Multi-year (my) daily timeseries 2D array (time, sites) for one
-            variable (daily total GHI, daily mean temp, etc...).
-        my_time_index : pd.datetimeindex
-            Daily datetime index corresponding to the rows in my_arr
-        time_masks : dict
-            Lookup of boolean masks for a daily timeseries keyed by month and
-            year integer.
-        """
-
-        if len(my_arr.shape) < 2:
-            raise TypeError('Input array is 1D, needs to be 2D (time, sites).')
-
-        self._my_arr = deepcopy(my_arr)
-        self._my_time_index = deepcopy(my_time_index)
-        self._years = sorted(self._my_time_index.year.unique())
-
-        self._my_arr, self._my_time_index = Tmy.drop_leap(
-            self._my_arr, self._my_time_index
-        )
-
-        if time_masks is None:
-            self._time_masks = Tmy._make_time_masks(self._my_time_index)
-        else:
-            masks = deepcopy(time_masks)
-            ti_temp = deepcopy(my_time_index)
-            for k, v in masks.items():
-                masks[k], _ = Tmy.drop_leap(v, ti_temp)
-            self._time_masks = masks
-
-        self._cdf = self._sort_monthly(self._my_arr)
-
-        self._lt_cdf = self._make_lt_cdf(self._cdf)
-
-        self._cdf_frac, self._lt_frac, self._interp_frac = self._cdf_fracs()
-
-        self._fs_all = self._fs_stat()
-
-    @property
-    def time_index(self):
-        """Get the multi-year daily non-leap time index.
-
-        Returns
-        -------
-        time_index : pd.datetimeindex
-            Daily datetime index corresponding to the rows in cdf.
-        """
-        return self._my_time_index
-
-    @property
-    def cdf(self):
-        """Get the multi-year cdf array with daily values sorted within each
-        year and month.
-
-        Returns
-        -------
-        cdf : np.ndarray
-            Multi-year daily timeseries array with sorted values
-            resetting each month.
-        """
-        return self._cdf
-
-    @property
-    def lt_cdf(self):
-        """Get the multi-year long term sorted array (values are sorted within
-        monthly masks across all years).
-
-        Returns
-        -------
-        lt_cdf : np.ndarray
-            Array with the same shape as arr but with monthly values sorted
-            across all years.
-        """
-        return self._lt_cdf
-
-    @property
-    def years(self):
-        """Get the list of years included in this multi-year cdf.
-
-        Returns
-        -------
-        years : list
-            List of years included in this cdf analysis.
-        """
-        return self._years
-
-    @property
-    def fs_all(self):
-        """Get all FS statistics for all months/years/sites.
-
-        Returns
-        -------
-        fs : dict
-            Dictionary with month keys. Each dict value is a (y, n) array where
-            y is years and n is sites. Each array entry is the FS metric.
-        """
-        return self._fs_all
-
-    @staticmethod
-    def _cumulative_sum_monthly(arr, time_index):
-        """Calculate the montly CDF of the data array.
-
-        Parameters
-        ----------
-        arr : np.ndarray
-            Timeseries array (time, sites) for one variable
-            (instantaneous data values).
-        time_index : pd.datetimeindex
-            Datetime index corresponding to the rows in arr.
-
-        Returns
-        -------
-        arr : np.ndarray
-            Timeseries array with cumulative values resetting each month.
-        """
-        years = time_index.year.unique()
-        for y in years:
-            year_mask = time_index.year == y
-            for m in range(1, 13):
-                mask = year_mask & (time_index.month == m)
-                arr[mask] = np.cumsum(arr[mask], axis=0)
-
-        return arr
-
-    def _sort_monthly(self, arr):
-        """Sort daily values within each month within each year.
-
-        Parameters
-        ----------
-        arr : np.ndarray
-            Timeseries array (time, sites) for one variable
-            (instantaneous or daily values).
-
-        Returns
-        -------
-        arr : np.ndarray
-            Timeseries array with sorted values resetting each month.
-        """
-
-        for y in self._years:
-            year_mask = self._time_masks[y]
-            for m in range(1, 13):
-                mask = year_mask & self._time_masks[m]
-                arr[mask] = np.sort(arr[mask], axis=0)
-
-        return arr
-
-    @staticmethod
-    def _resample_daily_max(arr, time_index):
-        """Convert a timeseries array to daily maximum data points.
-
-        Parameters
-        ----------
-        arr : np.ndarray
-            Timeseries array (time, sites) for one variable.
-        time_index : pd.datetimeindex
-            Datetime index corresponding to the rows in arr.
-
-        Returns
-        -------
-        arr : np.ndarray
-            Daily timeseries array where each value is the maximum value in
-            the given day.
-        time_index : pd.datetimeindex
-            Daily datetime index corresponding to the rows in arr.
-        """
-        df = pd.DataFrame(arr, index=time_index)
-        df = df.resample('1D').max()
-
-        return df.values, df.index
-
-    def _make_lt_cdf(self, arr):
-        """Make a long term value sorted array where values within month masks
-        are sorted across multi-years.
-
-        Parameters
-        ----------
-        arr : np.ndarray
-            Timeseries array with cumulative values.
-
-        Returns
-        -------
-        lt_cdf : np.ndarray
-            Array with the same shape as arr but with monthly values sorted
-            across all years.
-        """
-        lt_cdf = np.zeros_like(arr)
-        for m in range(1, 13):
-            mask = self._time_masks[m]
-            lt_cdf[mask, :] = np.sort(arr[mask], axis=0)
-        return lt_cdf
-
-    def _cdf_fracs(self):
-        """Make the fractional arrays for the y-axis of a CDF.
-
-        Returns
-        -------
-        cdf_frac : np.ndarray
-            (t, ) array of the cumulative summation fraction (0 to 1)
-            corresponding to the individual month and year for CDFs.
-        lt_frac : np.ndarray
-            (t, ) array of the cumulative summation fraction (0 to 1)
-            corresponding to months over multiple years in the long term cdf.
-        interp_frac : np.ndarray
-            (t, n) array of the cumulative summation fraction (0 to 1)
-            corresponding to the long term multi-year CDF. This is
-            the cdf fraction y-projected onto the long term CDF so that a
-            broadcasted subtraction can be performed.
-        """
-
-        cdf_frac = np.zeros((len(self._cdf), self._cdf.shape[1]))
-        lt_frac = np.zeros((len(self._lt_cdf), self._lt_cdf.shape[1]))
-        for m in range(1, 13):
-            mask = self._time_masks[m]
-            lt_frac[mask, :] = np.expand_dims(
-                np.linspace(0, 1, mask.sum()), axis=1
-            )
-            for y in self._years:
-                mask = self._time_masks[y] & self._time_masks[m]
-                cdf_frac[mask, :] = np.expand_dims(
-                    np.linspace(0, 1, mask.sum()), axis=1
-                )
-
-        interp_frac = np.zeros(self._cdf.shape)
-        for n in range(self._cdf.shape[1]):
-            for m in range(1, 13):
-                for y in self._years:
-                    lt_mask = self._time_masks[m]
-                    mask = self._time_masks[y] & self._time_masks[m]
-
-                    interp_frac[mask, n] = np.interp(
-                        self._cdf[mask, n],
-                        self._lt_cdf[lt_mask, n],
-                        lt_frac[lt_mask, n],
-                    )
-
-        return cdf_frac, lt_frac, interp_frac
-
-    def _fs_stat(self):
-        """Finkelstein-Schafer metric comparing the test cdf to a baseline
-
-        Returns
-        -------
-        fs : dict
-            Dictionary with month keys. Each dict value is a (y, n) array where
-            y is years and n is sites. Each array entry is the FS metric.
-        """
-
-        diff_cdf = np.abs(self._cdf_frac - self._interp_frac)
-
-        fs_arr = np.zeros((len(self._years), self._cdf.shape[1]))
-        fs = {m: deepcopy(fs_arr) for m in range(1, 13)}
-        for i, y in enumerate(self._years):
-            for m in range(1, 13):
-                mask = self._time_masks[y] & self._time_masks[m]
-                fs[m][i, :] = diff_cdf[mask, :].mean(axis=0)
-
-        return fs
-
-    def _best_fs_year(self):
-        """Select single best TMY year for each month based on the FS statistic
-
-        Returns
-        -------
-        years : np.ndarray
-            Array of best TMY years for every month for every site.
-            Shape is (months, sites).
-        fs : np.ndarray
-            Array of min FS statistic for each month for every site
-            corresponding to the year selected. Shape is (months, sites).
-        """
-
-        years = None
-        fs = None
-
-        for m, ranks in self._fs_all.items():
-            if years is None:
-                years = np.zeros((12, ranks.shape[1]), dtype=np.uint16)
-                fs = np.zeros((12, ranks.shape[1]))
-
-            i = m - 1
-            years[i, :] = int(self.years[0]) + np.argmin(ranks, axis=0)
-            fs[i, :] = np.min(ranks, axis=0)
-
-        return years, fs
-
-    def plot_tmy_selection(
-        self,
-        month=1,
-        site=0,
-        fig_size=(12, 9),
-        fout=None,
-        xlabel='Cumulative Value',
-        ylabel='CDF',
-        plot_years=None,
-    ):
-        """Plot a single site's TMY CDFs for one month.
-
-        Parameters
-        ----------
-        month : int
-            Month number to plot (1-indexed)
-        site : int
-            Site index to plot (0-indexed)
-        fig_size : tuple
-            Figure size (height, width)
-        fout : str | None
-            Filepath to save image to.
-        xlabel : str
-            Xlabel string (cumulative data value).
-        ylabel : str
-            Ylabel string (CDF fraction).
-        plot_years : list | None
-            Optional set of years to plot (makes the plot less busy)
-        """
-
-        tmy_years, _ = self._best_fs_year()
-
-        import matplotlib.pyplot as plt
-
-        fig = plt.figure(figsize=fig_size)
-        ax = fig.add_subplot(111)
-
-        if plot_years is None:
-            plot_years = self.years
-        for y in plot_years:
-            mask = self._time_masks[y] & self._time_masks[month]
-            ax.plot(self.cdf[mask, site], self._cdf_frac[mask, site], '--')
-
-        lt_mask = self._time_masks[month]
-        ax.plot(self.lt_cdf[lt_mask, site], self._lt_frac[lt_mask, site], 'b-')
-
-        tmy_year = tmy_years[(month - 1), site]
-        mask = self._time_masks[tmy_year] & self._time_masks[month]
-        ax.plot(self.cdf[mask, site], self._interp_frac[mask, site], 'rx')
-        ax.plot(self.cdf[mask, site], self._cdf_frac[mask, site], 'r-x')
-
-        legend = [
-            *plot_years,
-            'Long Term CDF',
-            'Interpolated',
-            'Best FS ({})'.format(tmy_year),
-        ]
-        plt.legend(legend)
-        ax.set_title('TMY CDFs for Month {} and Site {}'.format(month, site))
-        ax.set_xlabel(xlabel)
-        ax.set_ylabel(ylabel)
-
-        if fout is not None:
-            fig.savefig(fout, bbox_inches='tight')
-        else:
-            plt.show()
-
-        plt.close()
 
 
 class Tmy:
@@ -604,7 +229,7 @@ class Tmy:
                 with FileSystem(fpath) as f, Resource(f, unscale=False) as res:
                     ti = res.time_index
 
-                temp = self.drop_leap(temp, ti)[0]
+                temp = drop_leap(temp, ti)[0]
 
             arr[mask, :] = temp
 
@@ -644,29 +269,6 @@ class Tmy:
         arr = self._resample_arr_daily(arr, self.my_time_index, fun)
 
         return arr
-
-    @staticmethod
-    def _make_time_masks(time_index):
-        """Make a time index mask lookup dict.
-
-        Parameters
-        ----------
-        time_index : pd.datetimeindex
-            Time index to mask.
-
-        Returns
-        -------
-        masks : dict
-            Lookup of boolean masks keyed by month and year integer.
-        """
-        masks = {}
-        years = time_index.year.unique()
-        months = time_index.month.unique()
-        for y in years:
-            masks[y] = time_index.year == y
-        for m in months:
-            masks[m] = time_index.month == m
-        return masks
 
     def _init_daily_ghi_temp(self):
         """Initialize daily total ghi and daily mean air temperature
@@ -753,7 +355,7 @@ class Tmy:
             )
             self._time_index += datetime.timedelta(minutes=30)
             if len(self._time_index) != 8760:
-                _, self._time_index = self.drop_leap(
+                _, self._time_index = drop_leap(
                     np.zeros((len(self._time_index), 1)), self._time_index
                 )
         return self._time_index
@@ -768,6 +370,8 @@ class Tmy:
             Array of best TMY years for every month for every site.
             Shape is (months, sites).
         """
+        if self._tmy_years_short is None:
+            self._tmy_years_short = self.calculate_tmy_years()
         return self._tmy_years_short
 
     @property
@@ -792,7 +396,7 @@ class Tmy:
             Lookup of boolean masks keyed by month and year integer.
         """
         if self._time_masks is None:
-            self._time_masks = self._make_time_masks(self.my_time_index)
+            self._time_masks = make_time_masks(self.my_time_index)
         return self._time_masks
 
     @property
@@ -806,9 +410,7 @@ class Tmy:
             year integer.
         """
         if self._daily_time_masks is None:
-            self._daily_time_masks = self._make_time_masks(
-                self.my_daily_time_index
-            )
+            self._daily_time_masks = make_time_masks(self.my_daily_time_index)
         return self._daily_time_masks
 
     @property
@@ -827,32 +429,6 @@ class Tmy:
 
             self._meta = meta.iloc[self._site_slice, :]
         return self._meta
-
-    @staticmethod
-    def drop_leap(arr, time_index):
-        """Make 365-day timeseries (TMY does not have leap days).
-
-        Parameters
-        ----------
-        arr : np.ndarray
-            Timeseries array (time, sites) for one variable.
-        time_index : pd.datetimeindex
-            Datetime index corresponding to the rows in arr. May have leap days
-
-        Returns
-        -------
-        arr : np.ndarray
-            Timeseries array (time, sites) for one variable, n_rows is a
-            multiple of 8760.
-        time_index : pd.datetimeindex
-            Datetime index corresponding to the rows in arr, without leap days.
-        """
-        if len(arr) % 8760 != 0:
-            leap_day = (time_index.month == 2) & (time_index.day == 29)
-            if any(leap_day):
-                arr = arr[~leap_day]
-                time_index = time_index[~leap_day]
-        return arr, time_index
 
     @property
     def years(self):
@@ -1193,7 +769,7 @@ class Tmy:
             with FileSystem(fpath) as f, Handler(f, unscale=unscale) as res:
                 ti = res.time_index
                 temp = res[dset, :, self._site_slice]
-                temp, ti = self.drop_leap(temp, ti)
+                temp, ti = drop_leap(temp, ti)
 
             if masks is None:
                 masks = {m: (ti.month == m) for m in range(1, 13)}
@@ -1234,12 +810,14 @@ class Tmy:
             Array of best TMY years for every month for every site.
             Shape is (months, sites).
         """
-
+        logger.info(
+            'Calculating best TMY years for every month, for '
+            f'{len(self.meta)} sites.'
+        )
         ws = self.get_weighted_fs()
         tmy_years_5, _ = self.select_fs_years(ws)
         tmy_years_5, _ = self.sort_years_mm(tmy_years_5)
         tmy_years, _, _ = self.persistence_filter(tmy_years_5)
-        self._tmy_years_short = tmy_years
         return tmy_years
 
     def get_tmy_timeseries(self, dset, unscale=True):
@@ -1264,717 +842,6 @@ class Tmy:
             return self.tmy_years_short
         if dset in ('tmy_year_long', 'tmy_year'):
             return self.tmy_years_long
-        tmy_years = self.calculate_tmy_years()
-        data = self._make_tmy_timeseries(dset, tmy_years, unscale=unscale)
-        return data
-
-
-class TmyRunner:
-    """Class to handle running TMY, collecting outs, and writing to files."""
-
-    WEIGHTS: ClassVar = {
-        'TMY': {
-            'max_air_temperature': 0.05,
-            'min_air_temperature': 0.05,
-            'mean_air_temperature': 0.1,
-            'max_dew_point': 0.05,
-            'min_dew_point': 0.05,
-            'mean_dew_point': 0.1,
-            'max_wind_speed': 0.05,
-            'mean_wind_speed': 0.05,
-            'sum_dni': 0.25,
-            'sum_ghi': 0.25,
-        },
-        'TDY': {'sum_dni': 1.0},
-        'TGY': {'sum_ghi': 1.0},
-    }
-
-    def __init__(
-        self,
-        nsrdb_base_fp,
-        years,
-        weights,
-        sites_per_worker=100,
-        n_nodes=1,
-        node_index=0,
-        site_slice=None,
-        out_dir='/tmp/scratch/tmy/',
-        fn_out='tmy.h5',
-        supplemental_fp=None,
-        var_meta=None,
-    ):
-        """
-        Parameters
-        ----------
-        nsrdb_base_fp : str
-            Base nsrdb filepath to retrieve annual files from. Must include
-            a single {} format option for the year. Can include * for an
-            NSRDB multi file source.
-        years : iterable
-            Iterable of years to include in the TMY calculation.
-        weights : dict
-            Lookup of {dset: weight} where dset is a variable h5 dset name
-            and weight is a fractional TMY weighting. All weights must
-            sum to 1.0
-        sites_per_worker : int
-            Number of sites to run at once (sites per core/worker).
-        n_nodes : int
-            Number of nodes being run.
-        node_index : int
-            Index of this node job.
-        site_slice : slice
-            Sites to consider in the GLOBAL TMY run. If multiple jobs are being
-            run, the site slice should be the same for all jobs, and slices the
-            full spatial extent meta data.
-        out_dir : str
-            Directory to dump temporary output files.
-        fn_out : str
-            Final output filename.
-        supplemental_fp : None | dict
-            Supplemental data base filepaths including {} for year for
-            uncommon dataset inputs to the TMY calculation. For example:
-            {'poa': '/projects/pxs/poa_h5_dir/poa_out_{}.h5'}
-        var_meta : str
-            CSV filepath containing meta data for all NSRDB variables.
-            Defaults to the NSRDB var meta csv in git repo.
-        """
-
-        logger.info('Initializing TMY runner for years: {}'.format(years))
-        logger.info('TMY weights: {}'.format(weights))
-
-        self._nsrdb_base_fp = nsrdb_base_fp
-        self._years = years
-        self._weights = weights
-
-        self._sites_per_worker = sites_per_worker
-        self._n_nodes = n_nodes
-        self._node_index = node_index
-        self._site_chunks = None
-        self._site_chunks_index = None
-
-        self._site_slice = (
-            slice(None)
-            if site_slice is None
-            else slice(*site_slice)
-            if isinstance(site_slice, list)
-            else site_slice
+        return self._make_tmy_timeseries(
+            dset, self.tmy_years_short, unscale=unscale
         )
-        self._meta = None
-        self._dsets = None
-
-        self._out_dir = out_dir
-        self._fn_out = fn_out
-        self._final_fpath = os.path.join(self._out_dir, self._fn_out)
-
-        self._supplemental_fp = supplemental_fp
-        self._var_meta = var_meta
-
-        if not os.path.exists(self._out_dir):
-            os.makedirs(self._out_dir)
-
-        self._tmy_obj = Tmy(
-            self._nsrdb_base_fp,
-            self._years,
-            self._weights,
-            site_slice=slice(0, 1),
-            supplemental_fp=supplemental_fp,
-        )
-
-        out = self._setup_job_chunks(
-            self.meta,
-            self._sites_per_worker,
-            self._n_nodes,
-            self._node_index,
-            self._out_dir,
-        )
-        self._site_chunks, self._site_chunks_index, self._f_out_chunks = out
-
-        logger.info('Node meta data is: \n{}'.format(self.meta))
-        logger.info(
-            'Node index {} with n_nodes {} running site chunks: '
-            '{} ... {}'.format(
-                node_index,
-                n_nodes,
-                str(self._site_chunks)[:100],
-                str(self._site_chunks)[-100:],
-            )
-        )
-        logger.info(
-            'Node index {} with n_nodes {} running site chunks '
-            'indices: {} ... {}'.format(
-                node_index,
-                n_nodes,
-                str(self._site_chunks_index)[:100],
-                str(self._site_chunks_index)[-100:],
-            )
-        )
-        logger.info(
-            'Node index {} with n_nodes {} running fout chunks: '
-            '{} ... {}'.format(
-                node_index,
-                n_nodes,
-                str(self._f_out_chunks)[:100],
-                str(self._f_out_chunks)[-100:],
-            )
-        )
-
-    @staticmethod
-    def _setup_job_chunks(
-        meta, sites_per_worker, n_nodes, node_index, out_dir
-    ):
-        """Setup chunks and file names for a multi-chunk multi-node job.
-
-        Parameters
-        ----------
-        meta : pd.DataFrame
-            FULL NSRDB meta data.
-        sites_per_worker : int
-            Number of sites to run at once (sites per core/worker).
-        n_nodes : int
-            Number of nodes being run.
-        node_index : int
-            Index of this node job (if a multi node job is being run).
-        out_dir : str
-            Directory to dump temporary output files.
-
-        Returns
-        -------
-        site_chunks : list
-             List of slices setting the site chunks to be run by this job.
-        site_chunks_index : list
-            List of integers setting the site chunk indices to be run by
-            this job.
-        f_out_chunks : dict
-            Dictionary of file output paths keyed by the site chunk indices.
-        """
-
-        arr = meta.index.values
-        tmp = np.array_split(arr, np.ceil(len(arr) / sites_per_worker))
-        site_chunks = [slice(x.min(), x.max() + 1) for x in tmp]
-        site_chunks_index = list(range(len(site_chunks)))
-
-        site_chunks = np.array_split(np.array(site_chunks), n_nodes)[
-            node_index
-        ].tolist()
-        site_chunks_index = np.array_split(
-            np.array(site_chunks_index), n_nodes
-        )[node_index].tolist()
-
-        f_out_chunks = {}
-        chunk_dir = os.path.join(out_dir, 'chunks/')
-        if not os.path.exists(chunk_dir):
-            os.makedirs(chunk_dir)
-        for ichunk in site_chunks_index:
-            f_out = os.path.join(chunk_dir, 'temp_out_{}.h5'.format(ichunk))
-            f_out_chunks[ichunk] = f_out
-
-        return site_chunks, site_chunks_index, f_out_chunks
-
-    @property
-    def meta(self):
-        """Get the full NSRDB meta data."""
-        if self._meta is None:
-            fpath, Handler = self._tmy_obj._get_fpath('ghi', self._years[0])
-            with FileSystem(fpath) as f, Handler(f) as res:
-                self._meta = res.meta.iloc[self._site_slice, :]
-
-        return self._meta
-
-    @property
-    def dsets(self):
-        """Get the NSRDB datasets excluding meta and time index."""
-        if self._dsets is None:
-            fpath, Handler = self._tmy_obj._get_fpath('ghi', self._years[0])
-            with FileSystem(fpath) as f, Handler(f) as res:
-                self._dsets = []
-                for d in res.dsets:
-                    if res.shapes[d] == res.shape:
-                        self._dsets.append(d)
-
-            if self._supplemental_fp is not None:
-                self._dsets += list(self._supplemental_fp.keys())
-
-            self._dsets.append('tmy_year')
-            self._dsets.append('tmy_year_short')
-            self._dsets = list(set(self._dsets))
-
-        return self._dsets
-
-    @property
-    def site_chunks(self):
-        """Get a list of site chunk slices to parallelize across"""
-        return self._site_chunks
-
-    @staticmethod
-    def get_dset_attrs(dsets, var_meta=None):
-        """Get output file dataset attributes for a set of datasets.
-
-        Parameters
-        ----------
-        dsets : list
-            List of dataset / variable names.
-        var_meta : str | pd.DataFrame | None
-            CSV file or dataframe containing meta data for all NSRDB variables.
-            Defaults to the NSRDB var meta csv in git repo.
-
-        Returns
-        -------
-        attrs : dict
-            Dictionary of dataset attributes keyed by dset name.
-        chunks : dict
-            Dictionary of chunk tuples keyed by dset name.
-        dtypes : dict
-            dictionary of numpy datatypes keyed by dset name.
-        """
-
-        attrs = {}
-        chunks = {}
-        dtypes = {}
-
-        for dset in dsets:
-            var_obj = VarFactory.get_base_handler(dset, var_meta=var_meta)
-            attrs[dset] = var_obj.attrs
-            chunks[dset] = var_obj.chunks
-            dtypes[dset] = var_obj.final_dtype
-
-            if 'units' in attrs[dset]:
-                attrs[dset]['psm_units'] = attrs[dset]['units']
-            if 'scale_factor' in attrs[dset]:
-                attrs[dset]['psm_scale_factor'] = attrs[dset]['scale_factor']
-
-        return attrs, chunks, dtypes
-
-    def _collect_chunk(self, f_out_chunk, site_slice, out):
-        """Add single chunk to final output."""
-        f_chunk_basename = os.path.basename(f_out_chunk)
-        with Resource(f_out_chunk, unscale=True) as chunk:
-            for dset in self.dsets:
-                try:
-                    data = chunk[dset]
-                except Exception as e:
-                    m = (
-                        f'Could not read file dataset "{dset}" from file '
-                        f'"{f_chunk_basename}". Received the following '
-                        f'exception: \n{e}'
-                    )
-                    logger.exception(m)
-                    raise e
-                else:
-                    out[dset, :, site_slice] = data
-        return out
-
-    def _collect(self, purge_chunks=False):
-        """Collect all chunked files into the final fout."""
-
-        status_file = os.path.join(self._out_dir, 'collect_status.txt')
-        status = self._pre_collect(status_file)
-        self._init_final_fout()
-
-        with Outputs(self._final_fpath, mode='a', unscale=True) as out:
-            for i, f_out_chunk in self._f_out_chunks.items():
-                site_slice = self.site_chunks[i]
-                f_chunk_basename = os.path.basename(f_out_chunk)
-
-                msg = f'Skipping file, already collected: {f_chunk_basename}'
-                if f_chunk_basename not in status:
-                    out = self._collect_chunk(
-                        f_out_chunk=f_out_chunk, site_slice=site_slice, out=out
-                    )
-                    with open(status_file, 'a') as f:
-                        f.write('{}\n'.format(os.path.basename(f_out_chunk)))
-                    msg = (
-                        f'Finished collecting #{i + 1} out of '
-                        f'{len(self._f_out_chunks)} for sites {site_slice} '
-                        f'from file {f_chunk_basename}'
-                    )
-                logger.info(msg)
-
-        if purge_chunks:
-            chunk_dir = os.path.dirname(
-                next(iter(self._f_out_chunks.values()))
-            )
-            logger.info('Purging chunk directory: {}'.format(chunk_dir))
-            shutil.rmtree(chunk_dir)
-
-    def _pre_collect(self, status_file):
-        """Check to see if all chunked files exist before running collect
-
-        Parameters
-        ----------
-        status_file : str
-            Filepath to status file with a line for each file that has been
-            collected.
-
-        Returns
-        -------
-        status : list
-            List of filenames that have already been collected.
-        """
-        missing = [
-            fp for fp in self._f_out_chunks.values() if not os.path.exists(fp)
-        ]
-        if any(missing):
-            emsg = 'Chunked file outputs are missing: {}'.format(missing)
-            logger.error(emsg)
-            raise FileNotFoundError(emsg)
-        msg = 'All chunked files found. Running collection.'
-        logger.info(msg)
-
-        status = []
-        if os.path.exists(status_file):
-            with open(status_file) as f:
-                status = f.readlines()
-                status = [s.strip('\n') for s in status]
-        return status
-
-    def _init_final_fout(self):
-        """Initialize the final output file."""
-        self._init_file(
-            self._final_fpath,
-            self.dsets,
-            self._tmy_obj.time_index,
-            self.meta,
-            var_meta=self._var_meta,
-        )
-
-    @staticmethod
-    def _init_file(f_out, dsets, time_index, meta, var_meta=None):
-        """Initialize an output file.
-
-        Parameters
-        ----------
-        f_out : str
-            File path to final .h5 file.
-        dsets : list
-            List of dataset names to initialize
-        time_index : pd.datetimeindex
-            Time index to init to file.
-        meta : pd.DataFrame
-            Meta data to init to file.
-        var_meta : str | pd.DataFrame | None
-            CSV file or dataframe containing meta data for all NSRDB variables.
-            Defaults to the NSRDB var meta csv in git repo.
-        """
-
-        if not os.path.isfile(f_out):
-            dsets_mod = [d for d in dsets if 'tmy_year' not in d]
-            attrs, chunks, dtypes = TmyRunner.get_dset_attrs(
-                dsets_mod, var_meta=var_meta
-            )
-            dsets_mod.append('tmy_year')
-            attrs['tmy_year'] = {
-                'units': 'selected_year',
-                'scale_factor': 1,
-                'psm_units': 'selected_year',
-                'psm_scale_factor': 1,
-            }
-            chunks['tmy_year'] = chunks['dni']
-            dtypes['tmy_year'] = np.uint16
-            Outputs.init_h5(
-                f_out, dsets_mod, attrs, chunks, dtypes, time_index, meta
-            )
-
-            with h5py.File(f_out, mode='a') as f:
-                d = 'tmy_year_short'
-                f.create_dataset(d, shape=(12, len(meta)), dtype=np.uint16)
-                f[d].attrs['units'] = 'selected_year'
-                f[d].attrs['scale_factor'] = 1
-                f[d].attrs['psm_units'] = 'selected_year'
-                f[d].attrs['psm_scale_factor'] = 1
-
-    @staticmethod
-    def _write_output(f_out, data_dict, time_index, meta, var_meta=None):
-        """Initialize and write an output file chunk.
-
-        Parameters
-        ----------
-        f_out : str
-            File path to final .h5 file.
-        data_dict : dict
-            {Dset: data_arr} dictionary
-        time_index : pd.datetimeindex
-            Time index to init to file.
-        meta : pd.DataFrame
-            Meta data to init to file.
-        var_meta : str | pd.DataFrame | None
-            CSV file or dataframe containing meta data for all NSRDB variables.
-            Defaults to the NSRDB var meta csv in git repo.
-        """
-
-        logger.debug('Saving TMY results to: {}'.format(f_out))
-
-        TmyRunner._init_file(
-            f_out, list(data_dict.keys()), time_index, meta, var_meta=var_meta
-        )
-
-        with Outputs(f_out, mode='a') as f:
-            for dset, arr in data_dict.items():
-                f[dset] = arr
-
-    @staticmethod
-    def _run_file(fp):
-        """Check whether to run tmy for a given output filepath based on
-        whether that file already exists and its file size."""
-        run = True
-        if os.path.exists(fp):
-            size = os.path.getsize(fp)
-            if size > 1e6:
-                run = False
-
-        return run
-
-    @staticmethod
-    def run_single(
-        nsrdb_base_fp,
-        years,
-        weights,
-        site_slice,
-        dsets,
-        f_out,
-        supplemental_fp=None,
-        var_meta=None,
-    ):
-        """Run TMY for a single site chunk (slice) and save to disk.
-
-        Parameters
-        ----------
-        nsrdb_base_fp : str
-            Base nsrdb filepath to retrieve annual files from. Must include
-            a single {} format option for the year. Can include * for an
-            NSRDB multi file source.
-        years : iterable
-            Iterable of years to include in the TMY calculation.
-        weights : dict
-            Lookup of {dset: weight} where dset is a variable h5 dset name
-            and weight is a fractional TMY weighting. All weights must
-            sum to 1.0
-        site_slice : slice
-            Sites to consider in this TMY chunk.
-        dsets : list
-            List of TMY datasets to make.
-        f_out : str
-            Filepath to save file for this chunk.
-        supplemental_fp : None | dict
-            Supplemental data base filepaths including {} for year for
-            uncommon dataset inputs to the TMY calculation. For example:
-            {'poa': '/projects/pxs/poa_h5_dir/poa_out_{}.h5'}
-        var_meta : str | pd.DataFrame | None
-            CSV file or dataframe containing meta data for all NSRDB variables.
-            Defaults to the NSRDB var meta csv in git repo.
-
-        Returns
-        -------
-        True
-        """
-
-        run = TmyRunner._run_file(f_out)
-
-        if run:
-            data_dict = {}
-            tmy = Tmy(
-                nsrdb_base_fp,
-                years,
-                weights,
-                site_slice,
-                supplemental_fp=supplemental_fp,
-            )
-            for dset in dsets:
-                data_dict[dset] = tmy.get_tmy_timeseries(dset)
-            TmyRunner._write_output(
-                f_out, data_dict, tmy.time_index, tmy.meta, var_meta=var_meta
-            )
-        else:
-            logger.info(
-                'Skipping chunk, f_out already exists: {}'.format(f_out)
-            )
-
-        return True
-
-    def _run_serial(self):
-        """Run serial tmy futures and save temp chunks to disk."""
-
-        logger.info('Running in serial.')
-        for i, site_slice in enumerate(self.site_chunks):
-            fi = self._site_chunks_index[i]
-            f_out = self._f_out_chunks[fi]
-            if self._run_file(f_out):
-                self.run_single(
-                    self._nsrdb_base_fp,
-                    self._years,
-                    self._weights,
-                    site_slice,
-                    self.dsets,
-                    f_out,
-                    supplemental_fp=self._supplemental_fp,
-                    var_meta=self._var_meta,
-                )
-            else:
-                logger.info('Skipping, already exists: {}'.format(f_out))
-
-            logger.info(
-                '{} out of {} TMY chunks completed.'.format(
-                    i + 1, len(self.site_chunks)
-                )
-            )
-
-    def _run_parallel(self):
-        """Run parallel tmy futures and save temp chunks to disk."""
-        futures = {}
-        loggers = ['nsrdb']
-
-        logger.info('Running in parallel.')
-        with SpawnProcessPool(loggers=loggers) as exe:
-            logger.info(
-                'Kicking off {} futures.'.format(len(self.site_chunks))
-            )
-            for i, site_slice in enumerate(self.site_chunks):
-                fi = self._site_chunks_index[i]
-                f_out = self._f_out_chunks[fi]
-                if self._run_file(f_out):
-                    future = exe.submit(
-                        self.run_single,
-                        self._nsrdb_base_fp,
-                        self._years,
-                        self._weights,
-                        site_slice,
-                        self.dsets,
-                        f_out,
-                        supplemental_fp=self._supplemental_fp,
-                        var_meta=self._var_meta,
-                    )
-                    futures[future] = i
-                else:
-                    logger.info('Skipping, already exists: {}'.format(f_out))
-
-            logger.info(
-                'Finished kicking off {} futures.'.format(len(futures))
-            )
-
-            for i, future in enumerate(as_completed(futures)):
-                if future.result():
-                    logger.info(
-                        '{} out of {} futures completed.'.format(
-                            i + 1, len(futures)
-                        )
-                    )
-                else:
-                    logger.warning('Future #{} failed!'.format(i + 1))
-
-    def _run(self):
-        """Run in serial or parallel depending on number of chunks."""
-        if len(self.site_chunks) > 1:
-            self._run_parallel()
-        else:
-            self._run_serial()
-
-    @classmethod
-    def tmy(
-        cls,
-        nsrdb_base_fp,
-        years,
-        out_dir,
-        fn_out,
-        tmy_type='tmy',
-        weights=None,
-        sites_per_worker=100,
-        n_nodes=1,
-        node_index=0,
-        log=True,
-        log_level='INFO',
-        log_file=None,
-        site_slice=None,
-        supplemental_fp=None,
-        var_meta=None,
-    ):
-        """Run the TMY. Option for custom weights. Select default weights for
-        TMY / TDY / TGY with `tmy_type` and `weights = None`"""
-
-        if log:
-            init_logger('nsrdb.tmy', log_level=log_level, log_file=log_file)
-
-        if weights is None:
-            weights = cls.WEIGHTS[tmy_type.upper()]
-
-        tmy = cls(
-            nsrdb_base_fp,
-            years,
-            weights,
-            sites_per_worker=sites_per_worker,
-            out_dir=out_dir,
-            fn_out=fn_out,
-            n_nodes=n_nodes,
-            node_index=node_index,
-            site_slice=site_slice,
-            supplemental_fp=supplemental_fp,
-            var_meta=var_meta,
-        )
-        tmy._run()
-
-    @classmethod
-    def collect(
-        cls,
-        nsrdb_base_fp,
-        years,
-        out_dir,
-        fn_out,
-        sites_per_worker=100,
-        site_slice=None,
-        supplemental_fp=None,
-        var_meta=None,
-        log=True,
-        log_level='INFO',
-        log_file=None,
-        purge_chunks=False,
-    ):
-        """Run TMY collection.
-
-        Parameters
-        ----------
-        nsrdb_base_fp : str
-            Base nsrdb filepath to retrieve annual files from. Must include
-            a single {} format option for the year. Can include * for an
-            NSRDB multi file source.
-        years : iterable
-            Iterable of years to include in the TMY calculation.
-        out_dir : str
-            Directory to dump temporary output files.
-        fn_out : str
-            Final output filename.
-        sites_per_worker : int
-            Number of sites to run at once (sites per core/worker). Used here
-            to determine size of chunks to collect. Needs to match the value
-            used during the initial call to `tmy()`.
-        site_slice : slice
-            Sites to consider in this TMY.
-        supplemental_fp : None | dict
-            Supplemental data base filepaths including {} for year for
-            uncommon dataset inputs to the TMY calculation. For example:
-            {'poa': '/projects/pxs/poa_h5_dir/poa_out_{}.h5'}
-        var_meta : str
-            CSV filepath containing meta data for all NSRDB variables.
-            Defaults to the NSRDB var meta csv in git repo.
-        log : bool
-            Whether to write to logger
-        log_level : str
-            Log level for log output
-        log_file : str | None
-            Optional log file to write log output
-        purge_chunks : bool
-            Whether to purge chunks after collection into single file.
-        """
-        if log:
-            init_logger('nsrdb.tmy', log_level=log_level, log_file=log_file)
-        tgy = cls(
-            nsrdb_base_fp,
-            years,
-            sites_per_worker=sites_per_worker,
-            weights={'sum_ghi': 1.0},
-            out_dir=out_dir,
-            fn_out=fn_out,
-            n_nodes=1,
-            node_index=0,
-            site_slice=site_slice,
-            supplemental_fp=supplemental_fp,
-            var_meta=var_meta,
-        )
-        tgy._collect(purge_chunks=purge_chunks)
